unifique los constructores de fantasmas, quitando el parametro innecesario, y subiendo un nivel las cosas comunes le agregue el atributo velocidadNativa, y el metodo para restablecerla modifique los tests para que compilen(xq los bloques piden una referencia a fantasma) modifique el programa para q se adapte a los nuevos constructores de fantasmas. logre que los puntos se muestren sin la coma

+++ src/vista/EstadoJuego.java
@@ -37,7 +37,7 @@ public class EstadoJuego implements Posicionable {
 	}
 	
 	public String puntos(){		
-		return Double.toString(juego.getPuntosAcumulados());
+		return Integer.toString((int)juego.getPuntosAcumulados());
 	}
 
 }
